Revert "LPS-73481 Move upgrade processes implemented after the release of GA3 to a new upgrade version"
This reverts commit f4fc96c69a2275cee5572b9b6a6b836cf156bfe3.
