+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_3.java
@@ -16,13 +16,19 @@ package com.liferay.portal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.ReleaseInfo;
+import com.liferay.portal.upgrade.v7_0_3.UpgradeContact;
+import com.liferay.portal.upgrade.v7_0_3.UpgradeEmailAddress;
+import com.liferay.portal.upgrade.v7_0_3.UpgradeExpando;
 import com.liferay.portal.upgrade.v7_0_3.UpgradeGroup;
+import com.liferay.portal.upgrade.v7_0_3.UpgradeMBMailingList;
 import com.liferay.portal.upgrade.v7_0_3.UpgradeMessageBoards;
 import com.liferay.portal.upgrade.v7_0_3.UpgradeModules;
 import com.liferay.portal.upgrade.v7_0_3.UpgradeOracle;
 import com.liferay.portal.upgrade.v7_0_3.UpgradeOrganization;
 import com.liferay.portal.upgrade.v7_0_3.UpgradeSQLServer;
+import com.liferay.portal.upgrade.v7_0_3.UpgradeSchema;
 import com.liferay.portal.upgrade.v7_0_3.UpgradeSybase;
+import com.liferay.portal.upgrade.v7_0_3.UpgradeUser;
 
 /**
  * @author Adolfo Pérez
@@ -36,13 +42,20 @@ public class UpgradeProcess_7_0_3 extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
+		upgrade(UpgradeSchema.class);
+
+		upgrade(UpgradeContact.class);
+		upgrade(UpgradeEmailAddress.class);
+		upgrade(UpgradeExpando.class);
 		upgrade(UpgradeGroup.class);
+		upgrade(UpgradeMBMailingList.class);
 		upgrade(UpgradeMessageBoards.class);
 		upgrade(UpgradeModules.class);
 		upgrade(UpgradeOrganization.class);
 		upgrade(UpgradeOracle.class);
 		upgrade(UpgradeSQLServer.class);
 		upgrade(UpgradeSybase.class);
+		upgrade(UpgradeUser.class);
 
 		clearIndexesCache();
 	}
